requirements.txt
# Core dependencies
gradio>=4.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0

# Video download
yt-dlp>=2025.0.0

# Audio processing
demucs @ git+https://github.com/facebookresearch/demucs#egg=demucs
scipy>=1.10.0
librosa>=0.10.0
audiostretchy>=1.3.0

# Speech recognition
whisperx @ git+https://github.com/m-bain/whisperx.git
openai-whisper

# AI/ML
openai>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0

# Text-to-Speech
TTS>=0.22.0

# TTS dependencies (required for proper functionality)
gruut>=2.2.0
gruut-ipa>=0.13.0
gruut_lang_en>=2.0.0
num2words>=0.5.10
pypinyin>=0.50.0
python-crfsuite>=0.9.7
umap-learn>=0.5.1

# Upload
bilibili_toolman>=1.0.7

# System utilities
psutil>=5.9.0

# Optional: For better compatibility
# Install PyTorch with CUDA support separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121